arrumando entity services repository e test

+++ src/application/Services/MovieService.py
@@ -1,6 +1,7 @@
 import uuid
 from src.application.DTOs.MovieDTO import MovieDTO
 from src.application.DTOs.RegionDTO import RegionDTO
+from src.application.Services.Interfaces.IMovieRegionTicketsPurchesedService import IMovieRegionTicketsPurchesedService
 from src.application.Services.Interfaces.IMovieService import IMovieService
 from src.application.Services.Interfaces.IMovieTheaterService import IMovieTheaterService
 from src.application.Services.Interfaces.IRegionService import IRegionService
@@ -12,10 +13,11 @@ from src.infradata.UtilityExternal.Interface.ICloudinaryUti import ICloudinaryUt
 
 
 class MovieService(IMovieService):
-    def __init__(self, movie_repository: IMovieRepository, movie_theater_service: IMovieTheaterService, region_service: IRegionService, cloudinary_util: ICloudinaryUti) -> None:
+    def __init__(self, movie_repository: IMovieRepository, movie_theater_service: IMovieTheaterService, region_service: IRegionService, movie_region_tickets_purchesed_service: IMovieRegionTicketsPurchesedService, cloudinary_util: ICloudinaryUti) -> None:
         self.__movie_repository = movie_repository
         self.__movie_theater_service = movie_theater_service
         self.__region_service = region_service
+        self.__movie_region_tickets_purchesed_service = movie_region_tickets_purchesed_service
         self.__cloudinary_util = cloudinary_util
 
     def get_movie_by_id_check_exist(self, movie_id: str) -> ResponseWrapper:
@@ -125,7 +127,11 @@ class MovieService(IMovieService):
         if not delete_movie_theater_result.IsSuccess:
             return delete_movie_theater_result
 
-        # criar aqui para deletar do 'MovieRegionTicketsPurchesed' tem que deletar a junção lá tbm
+        delete_movie_region_tickets_purchesed = self.__movie_region_tickets_purchesed_service.delete(
+            id_movie)
+
+        if not delete_movie_region_tickets_purchesed.IsSuccess:
+            return delete_movie_region_tickets_purchesed
 
         movie_delete_result = self.__movie_repository.delete(id_movie)
